requirements.txt
Flask==2.3.3
requests==2.31.0
loguru==0.7.3
python-dotenv==1.0.0